aufgabe2

+++ fakultaet_long.c
@@ -2,19 +2,19 @@
 #include <math.h>
 #include <stdlib.h>
 
-int main (int argc, char **argv)
+long main (long argc, char **argv)
 {
-	int factorial = 1;	
-	int x;
+	long factorial = 1;	
+	long x;
 	x= atof(argv[1]);
 
 	if (argc == 2){ 
 	
-        //calculate factorial   
-        for (int i = 1; i <= x; i++){
+        //calculate factorial 
+        for (long i = 1; i <= x; i++){
         factorial *= i; 
 }    	
-	printf("factorial of %i is %i\n", x, factorial);
+	printf("factorial of %ld is %ld\n", x, factorial);
 	printf("number is processed with %ld Bytes\n", sizeof(factorial));
 }
 	else{
@@ -23,5 +23,5 @@ int main (int argc, char **argv)
 	return 0;
 }
 
-// programm runs correct till 12! 
+// programm runs correct till 20!
